+++ src/main/web/report/finalist/finalist.js
@@ -172,6 +172,7 @@ const finalist_module = {}
         this.org = org;
         this.judgingGroup = judgingGroup;
         this.categoryScores = {};
+        this.weightedRank = {};
         // names of playoff brackets this team is competing in
         this.playoffDivisions = [];
         _teams[num] = this;
@@ -192,8 +193,8 @@ const finalist_module = {}
     function Category(name, numeric, overall) {
         let category_id;
         // find the next available id
-        for (category_id = 0; category_id < Number.MAX_VALUE
-            && _categories[category_id]; category_id = category_id + 1)
+        for (category_id = 0;category_id < Number.MAX_VALUE
+            && _categories[category_id];category_id = category_id + 1)
             ;
 
         if (category_id == Number.MAX_VALUE || category_id + 1 == Number.MAX_VALUE) {
@@ -599,6 +600,14 @@ const finalist_module = {}
 
     finalist_module.setCategoryScore = function(team, category, score) {
         team.categoryScores[category.catId] = score;
+    };
+
+    finalist_module.getWeightedRank = function(team, category) {
+        return team.weightedRank[category.catId];
+    };
+
+    finalist_module.setWeightedRank = function(team, category, rank) {
+        team.weightedRank[category.catId] = rank;
     };
 
     /**
@@ -1769,13 +1778,14 @@ const finalist_module = {}
                 throw new Error("Missing championship category");
             }
 
-            for (const [teamNumber, score] of Object.entries(data)) {
+            for (const [teamNumber, scoreData] of Object.entries(data)) {
                 const team = finalist_module.lookupTeam(teamNumber);
                 if (null == team) {
                     throw new Error("Cannot find team with " + teamNumber + " found in overall scores");
                 }
 
-                finalist_module.setCategoryScore(team, championship, score);
+                finalist_module.setCategoryScore(team, championship, scoreData.overallScore);
+                finalist_module.setWeightedRank(team, championship, scoreData.weightedRank);
             } // scores
         });
     };